+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Change Customize Columns items icon color (#692)
 -   Increase page size to 30 (#703)
 -   CP Workflow graph: Allow a higher min zoom to show more tasks
+-   CP Workflow graph: update layout for predict tasks and test tasks
 
 ### Fixed